requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.18.0
anyio==3.6.2
astunparse==1.6.3
async-lru==1.0.3
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
blis==0.7.9
boto3==1.26.14
botocore==1.29.14
bs4==0.0.1
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
confection==0.0.3
configobj==5.0.6
configparser==5.3.0
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
Cython==0.29.33
databases==0.6.2
decorator==5.1.1
docopt==0.6.2
etelemetry==0.3.0
fastapi==0.88.0
filelock==3.8.0
fire==0.5.0
fitz==0.0.1.dev2
flatbuffers==23.3.3
fonttools==4.38.0
frequency-controller==0.0.6
frozenlist==1.3.3
future==0.18.2
gast==0.4.0
gensim==4.3.1
google-api-core==2.11.0
google-api-python-client==2.71.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.1
greenlet==2.0.1
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
httplib2==0.21.0
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
keras==2.11.0
keyboard==0.13.5
kiwisolver==1.4.4
langcodes==3.3.0
libclang==15.0.6.1
looseversion==1.0.2
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.3
multidict==6.0.4
murmurhash==1.0.9
networkx==2.8.8
nibabel==4.0.2
nipype==1.8.5
nltk==3.8.1
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
objectpath==0.6.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
pathlib==1.0.1
pathy==0.8.1
pbr==5.11.0
Pillow==9.4.0
pip==22.0.4
pip-autoremove==0.10.0
pipreqs==0.4.11
ply==3.11
preshed==3.0.8
protobuf==3.19.6
prov==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pydot==1.4.2
PyDrive==1.3.1
PyMuPDF==1.21.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
pywin32==305
pywinauto==0.6.8
pyxnat==1.5
PyYAML==6.0
rdflib==6.2.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.9.3
seaborn==0.12.2
selectolax==0.3.12
setuptools==58.1.0
simplejson==3.18.0
six==1.16.0
sklearn==0.0.post1
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.41
srsly==2.4.5
starlette==0.22.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
thinc==8.1.5
threadpoolctl==3.1.0
toml==0.10.2
torequests==6.0.0
tqdm==4.64.1
traits==6.3.2
typer==0.7.0
typing_extensions==4.4.0
uniparser==3.0.2
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.20.0
wasabi==0.10.1
watchdog==2.2.1
watchdogs==1.9.10
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.15.0
yarg==0.1.9
yarl==1.8.2
